Refactor nonce into its own module
so that we can keep `FcLogin` relatively simple.
